% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_Correction.R
\name{apply_bh_correction}
\alias{apply_bh_correction}
\title{Apply Benjamini-Hochberg Correction}
\usage{
apply_bh_correction(data, pval_col = "p.value.fisher")
}
\arguments{
\item{data}{A data frame containing a column of p-values.}

\item{pval_col}{The name of the column containing p-values (default \code{"p.value.fisher"}).}
}
\value{
The input data frame with an added column \code{bh_adjusted_p}.
}
\description{
Adjusts p-values using the BH method.
}
