% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_fisher_pvalues}
\alias{calculate_fisher_pvalues}
\title{Calculate Fisher's Combined p-value}
\usage{
calculate_fisher_pvalues(data, pval_col = "p.value")
}
\arguments{
\item{data}{A data frame containing at least two columns: "Gene.names" and a column with p-values.}

\item{pval_col}{A string indicating the name of the column containing p-values. Default is \code{"p.value"}.}
}
\value{
A data frame with \code{Gene.names}, Fisher's combined p-value, and the number of peptides per gene.
}
\description{
This function takes a data frame containing p-values and computes Fisher's combined p-value
for each unique gene using the \code{sumlog()} function from the \code{metap} package.
}
\examples{
df <- data.frame(Gene.names = rep(c("Gene1", "Gene2"), each = 3),
                 p.value = c(0.01, 0.02, 0.05, 0.001, 0.02, 0.03))
calculate_fisher_pvalues(df, "p.value")

}
