% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_shift_compute.R
\name{compute_mean_shift}
\alias{compute_mean_shift}
\title{Compute Mean Shift for Each Protein}
\usage{
compute_mean_shift(peptide_data)
}
\arguments{
\item{peptide_data}{A data frame with \code{Gene.names} and \code{Mean_Shift} columns.}
}
\value{
A data frame with each \code{Gene.names} and its mean \code{Mean_Shift}.
}
\description{
Calculates the mean \code{Mean_Shift} for each unique protein.
}
